+++ cfdm/__init__.py
@@ -93,7 +93,7 @@ from .functions import (
     environment,
     log_level,
     rtol,
-    unique_domains,
+    unique_constructs,
     _log_level,
     _disable_logging,
     _reset_log_emergence_level,